requirements.txt
pandas>=2.0
numpy>=1.25
scipy>=1.11
statsmodels>=0.14
scikit-learn>=1.3
matplotlib>=3.8
python-dotenv>=1.0
requests>=2.31